extract create_job module
