+++ backend/conftest.py
@@ -10,6 +10,7 @@ import warnings
 from ianalyzer.elasticsearch import client_from_config
 from addcorpus.python_corpora.save_corpus import load_and_save_all_corpora
 from es import es_index as index, sync
+from indexing.create_job import create_indexing_job
 from django.conf import settings
 from django.contrib.auth.models import Group
 from addcorpus.models import Corpus
@@ -165,7 +166,7 @@ def _index_test_corpus(es_client: Elasticsearch, corpus_name: str):
 
     if not es_client.indices.exists(index=corpus.configuration.es_index):
         with warnings.catch_warnings():
-            job = index.create_indexing_job(corpus)
+            job = create_indexing_job(corpus)
             index.perform_indexing(job)
 
         # ES is "near real time", so give it a second before we start searching the index
