+++ backend/indexing/management/commands/index.py
@@ -5,7 +5,8 @@ from django.core.management import BaseCommand
 from addcorpus.python_corpora.load_corpus import load_corpus_definition
 from addcorpus.python_corpora.save_corpus import load_all_corpus_definitions
 from addcorpus.models import Corpus
-from es.es_index import perform_indexing, create_indexing_job
+from es.es_index import perform_indexing
+from indexing.create_job import create_indexing_job
 
 class Command(BaseCommand):
     help = '''
